Fixed Issue when defining an ingest AOI and the clearing it
Issue when defining an ingest AOI, clearing it, and then doing an
ingest without an AOI defined.  The previous AOI persists and defines
the ingest area.  Modified clearAoi function in drawFeaturesAdmin.js to
set the AppIngestTileAdmin.objIngestImage.aoi = “”.  This effectively
resets the AOI so that a user can then ingest the whole image without
clipping occurring.